requirements.txt
# pip install -r requirements.txt
torch == 1.8.0
numpy
pandas
transformers == 4.16.2
torchtext == 0.9.0
tqdm
scipy
gensim == 4.2.0
tensorboard
wandb
